requirements.txt
cffi==1.15.0
cryptography==36.0.2
pycparser==2.21
